Move Foobar into separate file for Scala 2.11
