+++ scanamo/src/test/scala/org/scanamo/generic/AutoDerivationTest.scala
@@ -63,17 +63,4 @@ class AutoDerivationTest extends AnyFunSuite with Matchers {
 
     result should ===(DynamoValue.fromString("foo"))
   }
-
 }
-
-object Foobar {
-  def fromString(s: String): Either[TypeCoercionError, Foobar] = s match {
-    case "foo" => Right(Foobar(()))
-    case _     => Left(TypeCoercionError(new RuntimeException(s"$s is not a foo")))
-  }
-
-  implicit val dynamoFormatFoo: DynamoFormat[Foobar] =
-    DynamoFormat.xmap[Foobar, String](fromString, (_: Foobar) => "foo")
-}
-
-case class Foobar(value: Unit)
